backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
boto3==1.34.34
python-dotenv==1.0.1
pydantic==2.6.1
pydantic-settings==2.2.1
python-multipart==0.0.9
websockets==12.0
httpx==0.27.0